refactor(tests): consolidate GeospatialIndex tests into test_grid.py
- Add reusable grid factory methods to test_grid_cases.py (vertex_minimal,
  unstructured_minimal, vertex_rectangular, etc.)
- Add shared fixtures to conftest.py wrapping GridCases methods
- Move unique GeospatialIndex tests to test_grid.py (test_index_build,
  test_index_repr, test_index_k_parameter, test_index_thin_sliver_cell)
- Remove redundant tests already covered by existing intersect tests
- Delete test_geospatial_index.py
